Reorganize `ActionsData` and `ActionData`
- Rename `ActionsData` into `ActionMap`.
- Rename `ActionData` into `Action`.
- Move `ActionMap`, `Action` and `ActionState` into `action_map` module.
